requirements.txt
langchain
langchain-groq
langchain-community
sqlalchemy
streamlit
python-dotenv
faker
